Added meet q page when users sign in through Codewhisperer (#4092)
* Added meet q page even when users sign in through cw

* added util

* switched q sign in
